Refactor: move classes to relevant packages
