+++ src/main/java/duke/command/Command.java
@@ -1,9 +1,9 @@
 package duke.command;
 
-import duke.DukeException;
 import duke.Storage;
-import duke.TaskList;
-import duke.Ui;
+import duke.exception.DukeException;
+import duke.task.TaskList;
+import duke.ui.Ui;
 
 /**
  * Super class of all Commands.
