+++ src/main/java/duke/command/UnmarkCommand.java
@@ -1,10 +1,10 @@
 package duke.command;
 
-import duke.DukeException;
 import duke.Storage;
-import duke.Task;
-import duke.TaskList;
-import duke.Ui;
+import duke.exception.DukeException;
+import duke.task.Task;
+import duke.task.TaskList;
+import duke.ui.Ui;
 
 /**
  * Represents command to unmark tasks as done.
