+++ src/main/java/duke/task/TaskType.java
@@ -1,4 +1,4 @@
-package duke;
+package duke.task;
 
 /**
  * Enum of current task types.
