+++ src/main/java/duke/ui/DialogBox.java
@@ -3,7 +3,6 @@ package duke.ui;
 import java.io.IOException;
 import java.util.Collections;
 
-import duke.Message;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.fxml.FXML;
