+++ src/main/java/duke/ui/Main.java
@@ -1,8 +1,8 @@
-package duke;
+package duke.ui;
 
 import java.io.IOException;
 
-import duke.ui.MainWindow;
+import duke.Duke;
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
