+++ src/main/java/duke/ui/MainWindow.java
@@ -1,7 +1,6 @@
 package duke.ui;
 
 import duke.Duke;
-import duke.Message;
 import javafx.application.Platform;
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
