+++ src/main/java/duke/ui/Message.java
@@ -1,4 +1,4 @@
-package duke;
+package duke.ui;
 
 /**
  * Represents the message with appropriate flags needed by the display.
